Atualizado Controller de clientes e produtos
